also list any builds from development NOT slated for release (SOFTWARE-2073)
that is, for each tag, list packages in osg-development that are not
currently in osg-testing.

SOFTWARE-2036 inspired.  The idea was, it will help us to catch stuff
that we meant to include in the release that we forgot to promote, etc.

It will also have the effect of reminding us of all the stuff that we
were consciously planning not to ship.  There is not an obvious way to
differentiate these two categories automatically, so humans will have to
make sense of it.



git-svn-id: https://vdt.cs.wisc.edu/svn/software/release-tools@21819 4e558342-562e-0410-864c-e07659590f8c