Change Python 2 scripts to Python 3
urllib2 is now urllib and the requests are subsequently changed, I don't know if they work yet. Also updated the strings with '.format(**locals())' to f-strings

(Squash)
Added directory and changed import.
osgtest/vendor directory is from osg-test repo. stringToVersion is a function from a vendor so it is not subject to change from Python 2 to 3.

Delete osgtest directory